+++ packages/client/react-base/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @crossmint/client-sdk-react-base
 
+## 1.0.7
+
+### Patch Changes
+
+- Updated dependencies [402001f]
+- Updated dependencies [e912d18]
+  - @crossmint/client-sdk-base@2.1.0
+  - @crossmint/wallets-sdk@0.21.0
+  - @crossmint/client-sdk-auth@1.2.51
+  - @crossmint/common-sdk-auth@1.0.73
+
+## 1.0.6
+
+### Patch Changes
+
+- 11ab4f6: Fix Phone Signer
+
 ## 1.0.5
 
 ### Patch Changes